Rename `parallel_frontend` module to `frontend_threads`
